feat: update dependencies, refactor modal imports, and implement new Login and Logout dialogs
