+++ package.json
@@ -25,7 +25,7 @@
     "lucide-react": "^0.515.0",
     "mongodb": "^6.21.0",
     "next": "16.0.10",
-    "next-auth": "5.0.0-beta.28",
+    "next-auth": "4.24.13",
     "next-themes": "^0.4.6",
     "radix-ui": "^1.4.3",
     "react": "^19.2.3",
@@ -33,7 +33,7 @@
     "react-hook-form": "^7.68.0",
     "react-hot-toast": "^2.6.0",
     "react-icons": "^5.5.0",
-    "shadcn": "^3.6.1",
+    "shadcn": "^3.6.2",
     "tailwind-merge": "^3.4.0",
     "zod": "^3.25.76"
   },
@@ -52,7 +52,7 @@
     "eslint": "^9.39.2",
     "eslint-config-next": "15.3.3",
     "husky": "^9.1.7",
-    "knip": "^5.74.0",
+    "knip": "^5.75.1",
     "lint-staged": "^16.2.7",
     "postcss": "^8.5.6",
     "prettier": "^3.7.4",
