+++ src/components/layouts/navbar.tsx
@@ -1,11 +1,11 @@
 'use client';
-import { Logo } from '@/components/ui/Logo';
+import { Logo } from '@/components/ui/logo';
 import { ModeSwitcher } from '@/components/ui/mode';
 import { UserDropdown } from '@/components/ui/user-dropdown';
 import { useRouter } from 'next/navigation';
 import { useSession } from 'next-auth/react';
 import { Button } from '@/components/ui/button';
-import { LoginDialog } from '@/components/modal/LoginDialog';
+import { LoginDialog } from '@/components/modal/login-dialog';
 import { useState } from 'react';
 
 export function Navbar() {
