+++ src/components/modal/login-dialog.tsx
@@ -7,7 +7,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog';
-import { Logo } from '@/components/ui/Logo';
+import { Logo } from '@/components/ui/logo';
 import { SignInButton } from '@/components/auth/sign-in-button';
 
 interface LoginDialogProps {
